fix: refactor common types
